+++ src/apps/actions.js
@@ -1,10 +1,39 @@
+import * as api from '../client-lib/api';
+
 export const NEW_BUNDLE = 'NEW_BUNDLE';
+export const UPDATE_BUNDLE = 'UPDATE_BUNDLE';
 export const NEW_ERROR = 'NEW_ERROR';
 export const FETCH_BUNDLE_ERROR = 'FETCH_BUNDLE_ERROR';
 
-export function newBundle(bundle, inputs) {
+export function refetchPathBundle() {
+    return (dispatch, getState) => {
+        let { juttleServiceHost, runMode } = getState();
+
+        if (runMode.path) {
+            let bundle;
+            api.getBundle(juttleServiceHost, runMode.path)
+            .then(res => {
+                bundle = res.bundle;
+                return api.describe(juttleServiceHost, bundle);
+            })
+            .then(inputs => {
+                dispatch({
+                    type: UPDATE_BUNDLE,
+                    bundle,
+                    inputs
+                });
+            })
+            .catch((err) => {
+                dispatch(newError(err));
+            });
+        }
+    }
+}
+
+export function newBundle(bundleId, bundle, inputs) {
     return {
         type: NEW_BUNDLE,
+        bundleId,
         bundle,
         inputs
     };