+++ src/apps/components/run.js
@@ -1,11 +1,12 @@
+import _ from 'underscore';
 import React from 'react';
 import { connect } from 'react-redux';
-import _ from 'underscore';
 
 import * as actions from '../actions';
 
 import Juttle from 'juttle-client-library';
 import JuttleViewer from './juttle-viewer';
+import ErrorView from './error-view';
 
 const ENTER_KEY = 13;
 
@@ -15,7 +16,6 @@ class RunApp extends React.Component {
         let client = new Juttle(this.props.juttleServiceHost);
         this.view = new client.View(this.refs.juttleViewLayout);
         this.inputs = new client.Input(this.refs.juttleInputsContainer);
-        this.errors = new client.Errors(this.refs.errorView);
 
         // subscribe to runtime errors
         this.view.on('error', this.runtimeError.bind(this, 'error'));
@@ -30,25 +30,37 @@ class RunApp extends React.Component {
 
     componentWillReceiveProps(nextProps) {
         let self = this;
-        if (nextProps.bundle !== this.props.bundle) {
-            this.errors.clear();
-            this.inputs.clear();
+        
+        let newBundle = nextProps.bundleId !== this.props.bundleId;
+        let bundleUpdated = newBundle || (nextProps.bundle !== this.props.bundle);
+        let inputDefsChanged = !_.isEqual(nextProps.inputs, this.props.inputs);
 
+        // if no bundle clear everything
+        if (!nextProps.bundle) {
+            this.inputs.clear();
+            this.view.clear();
+            return;
+        }
+
+        if (newBundle || inputDefsChanged) {
+            this.inputs.clear();
+            this.inputs.render(nextProps.bundle);
+        }
+
+        if (bundleUpdated) {
             this.view.clear()
             .then(() => {
-                if (nextProps.bundle) {
-                    this.inputs.render(nextProps.bundle);
-                    // if no inputs run view automagically
-                    if (nextProps.inputs.length === 0) {
-                        setTimeout(() => { self.runView() });
-                    }
+                // if not new bundle or new with no inputs run view automagically
+                if (!newBundle || nextProps.inputs.length === 0) {
+                    setTimeout(() => {
+                        this.inputs.getValues()
+                        .then(values => this.view.run(nextProps.bundle, values))
+                        .catch(err => { self.props.dispatch(actions.newError(err)); });
+                    });
                 }
-            })
+            });
         }
 
-        if (nextProps.error && !_.isEqual(nextProps.error, this.props.error)) {
-            this.errors.render(nextProps.error);
-        }
     }
 
     runtimeError = (type, err) => {
@@ -66,7 +78,11 @@ class RunApp extends React.Component {
     };
 
     handleRunClick = () => {
-        this.runView();
+        if (this.props.runMode.path) {
+            this.props.dispatch(actions.refetchPathBundle());
+        } else {
+            this.runView();
+        }
     };
 
     render() {
@@ -76,7 +92,7 @@ class RunApp extends React.Component {
                     <JuttleViewer bundle={this.props.bundle} />
                     <div ref="juttleSource"></div>
                     <div ref="juttleViewLayout"></div>
-                    <div ref="errorView"></div>
+                    <ErrorView error={this.props.error}/>
                 </div>
                 <div className="right-rail">
                     <div ref="juttleInputsContainer" onKeyDown={this._onInputContainerKeyDown}></div>
@@ -94,7 +110,9 @@ class RunApp extends React.Component {
 
 function select(state) {
     return {
+        runMode: state.runMode,
         error: state.bundleInfo.error,
+        bundleId: state.bundleInfo.bundleId,
         bundle: state.bundleInfo.bundle,
         inputs: state.bundleInfo.inputs,
         juttleServiceHost: state.juttleServiceHost